Bypass cg_trim() - rmdir(/sys/fs/cgroup/...) in Public Wearable(TW1) images based on Linux kernel 3.4
WARNING, HACKS AHEAD!

Tizen Public Wearable(TW1) images use Linux kernel 3.4, which does not satisfy the Systemd-219 requirement.

REQUIREMENTS:
        Linux kernel >= 3.7
        Linux kernel >= 3.8 for Smack support

It is observed that the following cgroup directories are created and never removed even when any oneshot service is executed.
/sys/fs/cgroup/systemd/*
/sys/fs/cgroup/memory/systemd/*

Also, Systemd(PID=1) is blocked in the following call stack while "systemctl start boot-animation.service".

$) cat /proc/1/wchan
wait_rcu_gp

(gdb) bt
.#0  0x4033c46c in rmdir () from /lib/libc.so.6
.#1  0x4004fbaa in cg_trim (controller=controller@entry=0x400c2576 <mask_names+18> "memory",
.    path=path@entry=0x415b2a18 "/system.slice/boot-animation.service", delete_root=delete_root@entry=true)
.    at src/shared/cgroup-util.c:570
.#2  0x40050cfc in cg_trim_everywhere (supported=supported@entry=11,
.    path=path@entry=0x415b2a18 "/system.slice/boot-animation.service", delete_root=<optimized out>) at src/shared/cgroup-util.c:1728
.#3  0x400417e8 in unit_destroy_cgroup_if_empty (u=u@entry=0x415d2d48) at src/core/cgroup.c:843
.#4  0x400883a4 in unit_notify (u=u@entry=0x415d2d48, os=UNIT_ACTIVATING, ns=UNIT_INACTIVE, reload_success=<optimized out>)
.    at src/core/unit.c:1811
.#5  0x40022a3c in service_set_state (s=s@entry=0x415d2d48, state=<optimized out>) at src/core/service.c:908
.#6  0x40022b62 in service_enter_dead (s=0x415d2d48, f=<optimized out>, allow_restart=<optimized out>) at src/core/service.c:1280
.#7  0x40026a10 in service_sigchld_event (u=0x415d2d48, pid=<optimized out>, code=<optimized out>, status=0)
.    at src/core/service.c:2484
.#8  0x400121c4 in manager_dispatch_sigchld (m=m@entry=0x41562158) at src/core/manager.c:1686
.#9  0x4001652a in manager_dispatch_signal_fd (source=<optimized out>, fd=<optimized out>, revents=<optimized out>,
.    userdata=0x41562158) at src/core/manager.c:1945
.#10 0x40081a30 in source_dispatch (s=s@entry=0x41562820) at src/libsystemd/sd-event/sd-event.c:2116
.#11 0x400825f4 in sd_event_dispatch (e=0x41562518) at src/libsystemd/sd-event/sd-event.c:2473
.#12 0x400172c6 in manager_loop (m=0x41562158) at src/core/manager.c:2064
.#13 0x4000ebc2 in main (argc=<optimized out>, argv=<optimized out>) at src/core/main.c:1756

This is a workaround patch to avoid this blocking issue by skipping rmdir(/sys/fs/cgroup/...) only in TW1 images.

+++ src/shared/cgroup-util.c
@@ -28,6 +28,7 @@
 #include <sys/stat.h>
 #include <sys/types.h>
 #include <ftw.h>
+#include <system/system_info.h>
 
 #include "cgroup-util.h"
 #include "log.h"
@@ -554,7 +555,8 @@ static int trim_cb(const char *path, const struct stat *sb, int typeflag, struct
 
 int cg_trim(const char *controller, const char *path, bool delete_root) {
         _cleanup_free_ char *fs = NULL;
-        int r = 0;
+        int r = 0, r_info = 0;
+        char *model_name = NULL;
 
         assert(path);
 
@@ -567,8 +569,23 @@ int cg_trim(const char *controller, const char *path, bool delete_root) {
                 r = errno ? -errno : -EIO;
 
         if (delete_root) {
-                if (rmdir(fs) < 0 && errno != ENOENT)
-                        return -errno;
+                bool rm = true;
+
+                r_info = system_info_get_platform_string("http://tizen.org/system/model_name", &model_name);
+
+                if(r_info == SYSTEM_INFO_ERROR_NONE && model_name)
+                {
+                        if(!strncmp(model_name, "TW1", 3))
+                                rm = false;
+                }
+
+                if(model_name)
+                        free(model_name);
+
+                if(rm) {
+                        if(rmdir(fs) < 0 && errno != ENOENT)
+                                r = -errno;
+                }
         }
 
         return r;
